requirements.txt
-i https://pypi.org/simple
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0a3
dj-database-url==0.5.0
django-filter==2.0.0
django-jet==1.0.8
django==2.1.3
djangorestframework==3.9.0
drf-yasg[validation]==1.11.0
flex==6.13.2 ; extra == 'validation'
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
jinja2==2.10
jsonfield==2.0.2
jsonpointer==1.14
jsonschema==3.0.0a3
markupsafe==1.1.0
more-itertools==4.3.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pyaml==17.12.1
pyrsistent==0.14.5
pyswagger==0.8.39
pytest-cov==2.6.0
pytest-django==3.4.4
pytest==4.0.0
python-decouple==3.1
pytz==2018.7
pyyaml==3.13
requests==2.20.1
rfc3987==1.3.8
ruamel.yaml==0.15.77
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1 ; extra == 'validation'
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
whitenoise==4.0
